rename florence2-ft to florence2 and rm florencev2
