+++ scripts/install.sh
@@ -1,6 +1,6 @@
 # Install all dependencies
 # Check if it is macos runtime or ubuntu
-# Assuming this is reun inside a python3 devcontainer
+# It assumes that you already have python installed
 set -e # exit on error
 
 curl https://bootstrap.pypa.io/get-pip.py -o get-pip.py
@@ -17,5 +17,4 @@ fi
 
 echo "Installing poetry dependencies"
 poetry install -E all
-# POETRY run install-dependencies
 echo "Installing poetry dependencies done 🎉"
