+++ tests/models/florence2/conftest.py
@@ -7,27 +7,27 @@ import cv2
 import pytest
 import numpy as np
 
+from vision_agent_tools.models.florence2 import Florence2
 from vision_agent_tools.shared_types import Florence2ModelName
-from vision_agent_tools.models.florence2_ft import Florence2Ft
 
 logging.basicConfig(level=logging.INFO)
 
 
 @pytest.fixture
 def large_model():
-    return Florence2Ft(Florence2ModelName.FLORENCE_2_LARGE)
+    return Florence2(Florence2ModelName.FLORENCE_2_LARGE)
 
 
 @pytest.fixture
 def small_model():
-    return Florence2Ft(Florence2ModelName.FLORENCE_2_BASE_FT)
+    return Florence2(Florence2ModelName.FLORENCE_2_BASE_FT)
 
 
 @pytest.fixture(scope="session")
 def shared_model():
     # do not fine-tune this model, if you do, remember to reset it to the base model
     # by calling shared_model.load_base()
-    return Florence2Ft(Florence2ModelName.FLORENCE_2_BASE_FT)
+    return Florence2(Florence2ModelName.FLORENCE_2_BASE_FT)
 
 
 @pytest.fixture
