+++ tests/models/florence2/data/results/caption_to_phrase_grounding_video_results.json
