+++ tests/models/florence2/data/results/ocr_results.json
