+++ tests/models/florence2/test_open_vocabulary_detection.py
@@ -4,9 +4,9 @@ from PIL import Image
 from vision_agent_tools.shared_types import PromptTask
 
 
-def test_detailed_caption(shared_model):
+def test_open_vocabulary_detection(shared_model):
     image_path = "tests/shared_data/images/cereal.jpg"
-    task = PromptTask.DETAILED_CAPTION
+    task = PromptTask.OPEN_VOCABULARY_DETECTION
     image = Image.open(image_path)
 
     payload = {
@@ -15,14 +15,14 @@ def test_detailed_caption(shared_model):
     }
     response = shared_model(**payload)
     assert response == [
-        {"text": "In this image we can see food items on the white surface."}
+        {"bboxes": [], "bboxes_labels": [], "polygons": [], "polygons_labels": []}
     ]
 
 
-def test_detailed_caption_ft(small_model, unzip_model):
+def test_open_vocabulary_detection_ft(small_model, unzip_model):
     image_path = "tests/shared_data/images/cereal.jpg"
-    task = PromptTask.DETAILED_CAPTION
-    model_zip_path = "tests/models/florence2_ft/data/models/od_checkpoint.zip"
+    task = PromptTask.OPEN_VOCABULARY_DETECTION
+    model_zip_path = "tests/models/florence2/data/models/od_checkpoint.zip"
     model_path = unzip_model(model_zip_path)
     image = Image.open(image_path)
 
@@ -35,5 +35,6 @@ def test_detailed_caption_ft(small_model, unzip_model):
         small_model(**payload)
         assert (
             exc.value
-            == "The task DETAILED_CAPTION is not supported yet if your are using a fine-tuned model."
+            == "The task OPEN_VOCABULARY_DETECTION is not supported yet if your are "
+            "using a fine-tuned model."
         )
