+++ tests/models/florence2/test_region_to_description.py
@@ -4,9 +4,9 @@ from PIL import Image
 from vision_agent_tools.shared_types import PromptTask
 
 
-def test_dense_region_caption(shared_model):
+def test_region_to_description(shared_model):
     image_path = "tests/shared_data/images/cereal.jpg"
-    task = PromptTask.DENSE_REGION_CAPTION
+    task = PromptTask.REGION_TO_DESCRIPTION
     image = Image.open(image_path)
 
     payload = {
@@ -14,13 +14,13 @@ def test_dense_region_caption(shared_model):
         "task": task,
     }
     response = shared_model(**payload)
-    assert response == [{"labels": [], "bboxes": []}]
+    assert response == [{"text": "doughnuts in various colors on a white surface"}]
 
 
-def test_dense_region_caption_ft(small_model, unzip_model):
+def test_region_to_description_ft(small_model, unzip_model):
     image_path = "tests/shared_data/images/cereal.jpg"
-    task = PromptTask.DENSE_REGION_CAPTION
-    model_zip_path = "tests/models/florence2_ft/data/models/od_checkpoint.zip"
+    task = PromptTask.REGION_TO_DESCRIPTION
+    model_zip_path = "tests/models/florence2/data/models/od_checkpoint.zip"
     model_path = unzip_model(model_zip_path)
     image = Image.open(image_path)
 
@@ -32,7 +32,6 @@ def test_dense_region_caption_ft(small_model, unzip_model):
     with pytest.raises(ValueError) as exc:
         small_model(**payload)
         assert (
-            exc.value
-            == "The task DENSE_REGION_CAPTION is not supported yet if your are using "
-            "a fine-tuned model."
+            exc.value == "The task REGION_TO_DESCRIPTION is not supported yet if "
+            "your are using a fine-tuned model."
         )
