+++ tests/shared_data/images/sheep_aerial.jpg
