+++ tests/tools/test_get_model.py
@@ -1,13 +1,14 @@
-from vision_agent_tools.models.model_registry import get_model_class
-from vision_agent_tools.models.florencev2 import Florencev2
 import pytest
+
+from vision_agent_tools.models.florence2 import Florence2
+from vision_agent_tools.models.model_registry import get_model_class
 
 
 def test_get_model_class_valid_model():
-    valid_model_name = "florencev2"
+    valid_model_name = "florence2"
     model_class = get_model_class(valid_model_name)
     model_instance = model_class()
-    assert isinstance(model_instance(), Florencev2)
+    assert isinstance(model_instance(), Florence2)
 
 
 def test_get_model_class_invalid_model():
